I´ve updated the script to perform CV with embedded feature selection.
Open issue:
I know that the function "resample" fits a model specified by Learner on a Task and calculates predictions and performance measures. But now this is done for different selected variables and not only for the "best" model/selection of variables.

As a solution, do I have to apply the function "resample" again but with a different task (-> with a new input dataset  that only contains the previously selected variables)?